+++ configuration.properties
@@ -1,5 +1,5 @@
 browser=chrome
-env=https://login1.nextbasecrm.com
-username=helpdesk1@cydeo.com
+env=https://login2.nextbasecrm.com
+username=helpdesk41@cydeo.com
 password=UserUser
 #comment using #hash sign
